pyproject.toml
[tool.poetry]
name = "llama-lifecycle"
version = "0.0.9a0"
description = ""
authors = ["Achilleas Athanasiou Fragkoulis <achilleasatha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs = "^1.5.3"
bumpver = "^2023.1129"
pydantic = "^2.6.0"
click = "^8.1.7"
pyyaml = "^6.0.1"
datasets = "^2.16.1"
sentencepiece = "^0.1.99"
torch = "^2.2.0"
transformers = "^4.37.2"
peft = "^0.8.2"
trl = "^0.7.11"
bitsandbytes = "^0.42.0"
accelerate = "^0.27.2"
protobuf = "^4.25.3"
setuptools = "^69.1.1"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
run_cli = "llama_lifecycle.cli.cli:run_cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bumpver]
current_version = "0.0.9a0"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit = true
commit_message = "bump version {old_version} -> {new_version}"
tag = true
push = true
bump_files = ["pyproject.toml"]

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{pep440_version}"'
]
"llama_lifecycle/__init__.py" = [
    '__version__ = "{pep440_version}"'
]

[tool.ruff]
line-length = 120
target-version = "py312"
